misc fixes
