+++ src/utils/use-isomorphic-layout-effect.ts
@@ -2,6 +2,6 @@
 // useLayoutEffect but does not trigger warning in server-side rendering
 import {useEffect, useLayoutEffect} from 'react';
 
-const useIsomorphicLayoutEffect = typeof window !== 'undefined' ? useLayoutEffect : useEffect;
+const useIsomorphicLayoutEffect = typeof document !== 'undefined' ? useLayoutEffect : useEffect;
 
 export default useIsomorphicLayoutEffect;
